Fix outdated generated mocks
