+++ pkg/lib/feature/verification/service_mock_test.go
@@ -5,107 +5,9 @@
 package verification
 
 import (
-	authenticator "github.com/authgear/authgear-server/pkg/lib/authn/authenticator"
-	identity "github.com/authgear/authgear-server/pkg/lib/authn/identity"
 	gomock "github.com/golang/mock/gomock"
 	reflect "reflect"
 )
-
-// MockIdentityService is a mock of IdentityService interface
-type MockIdentityService struct {
-	ctrl     *gomock.Controller
-	recorder *MockIdentityServiceMockRecorder
-}
-
-// MockIdentityServiceMockRecorder is the mock recorder for MockIdentityService
-type MockIdentityServiceMockRecorder struct {
-	mock *MockIdentityService
-}
-
-// NewMockIdentityService creates a new mock instance
-func NewMockIdentityService(ctrl *gomock.Controller) *MockIdentityService {
-	mock := &MockIdentityService{ctrl: ctrl}
-	mock.recorder = &MockIdentityServiceMockRecorder{mock}
-	return mock
-}
-
-// EXPECT returns an object that allows the caller to indicate expected use
-func (m *MockIdentityService) EXPECT() *MockIdentityServiceMockRecorder {
-	return m.recorder
-}
-
-// ListByUser mocks base method
-func (m *MockIdentityService) ListByUser(userID string) ([]*identity.Info, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ListByUser", userID)
-	ret0, _ := ret[0].([]*identity.Info)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// ListByUser indicates an expected call of ListByUser
-func (mr *MockIdentityServiceMockRecorder) ListByUser(userID interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ListByUser", reflect.TypeOf((*MockIdentityService)(nil).ListByUser), userID)
-}
-
-// MockAuthenticatorService is a mock of AuthenticatorService interface
-type MockAuthenticatorService struct {
-	ctrl     *gomock.Controller
-	recorder *MockAuthenticatorServiceMockRecorder
-}
-
-// MockAuthenticatorServiceMockRecorder is the mock recorder for MockAuthenticatorService
-type MockAuthenticatorServiceMockRecorder struct {
-	mock *MockAuthenticatorService
-}
-
-// NewMockAuthenticatorService creates a new mock instance
-func NewMockAuthenticatorService(ctrl *gomock.Controller) *MockAuthenticatorService {
-	mock := &MockAuthenticatorService{ctrl: ctrl}
-	mock.recorder = &MockAuthenticatorServiceMockRecorder{mock}
-	return mock
-}
-
-// EXPECT returns an object that allows the caller to indicate expected use
-func (m *MockAuthenticatorService) EXPECT() *MockAuthenticatorServiceMockRecorder {
-	return m.recorder
-}
-
-// List mocks base method
-func (m *MockAuthenticatorService) List(userID string, filters ...authenticator.Filter) ([]*authenticator.Info, error) {
-	m.ctrl.T.Helper()
-	varargs := []interface{}{userID}
-	for _, a := range filters {
-		varargs = append(varargs, a)
-	}
-	ret := m.ctrl.Call(m, "List", varargs...)
-	ret0, _ := ret[0].([]*authenticator.Info)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// List indicates an expected call of List
-func (mr *MockAuthenticatorServiceMockRecorder) List(userID interface{}, filters ...interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{userID}, filters...)
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "List", reflect.TypeOf((*MockAuthenticatorService)(nil).List), varargs...)
-}
-
-// New mocks base method
-func (m *MockAuthenticatorService) New(spec *authenticator.Spec, secret string) (*authenticator.Info, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "New", spec, secret)
-	ret0, _ := ret[0].(*authenticator.Info)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// New indicates an expected call of New
-func (mr *MockAuthenticatorServiceMockRecorder) New(spec, secret interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "New", reflect.TypeOf((*MockAuthenticatorService)(nil).New), spec, secret)
-}
 
 // MockCodeStore is a mock of CodeStore interface
 type MockCodeStore struct {
